Redesign Space/Container interactions to simplify working with bitsets in Features
The idea is that a Space syncs its state of alive and enabled stuff
down into all the containers in its Features.
This only requires a tiny bit of wiring from the FeatureSet impl.
This way Feature implementations don't need to (and can't forget to)
borrow this stuff from the Space manually.