Add (rudimentary) resistance to bending to ropes
Currently this takes quite a bit of tuning to be stable.
A good set of parameters can be made the default of course,
but maybe there's something we can do to improve this
on the physics side. Extra damping on the angle corrections?
Restitution-like velocity step to prevent bouncing back and forth?